fixed readme made pip installable fixed imports tidied up a bit
